Rename .Common namespace to .Model, relocate files.  Add Model.Library project.
