--- src/Band.Personalize/Band.Personalize.Common.Test/Package.appxmanifest
@@ -1,45 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package
-  xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
-  xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest"
-  xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
-  IgnorableNamespaces="uap mp">
-
-  <Identity Name="e6b41e07-ed01-4e88-ad41-65f3958e0a8b"
-            Publisher="CN=butchern"
-            Version="1.0.0.0" />
-
-  <mp:PhoneIdentity PhoneProductId="e6b41e07-ed01-4e88-ad41-65f3958e0a8b" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
-
-  <Properties>
-    <DisplayName>Band.Personalize.Common.Test</DisplayName>
-    <PublisherDisplayName>butchern</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-  
-  <Dependencies>
-    <TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.0.0" MaxVersionTested="10.0.0.0" />
-  </Dependencies>
-
-  <Resources>
-    <Resource Language="x-generate" />
-  </Resources>
-  <Applications>
-    <Application Id="vstest.executionengine.universal.App" 
-        Executable="$targetnametoken$.exe"
-        EntryPoint="Band.Personalize.Common.Test.App">
-      <uap:VisualElements
-        DisplayName="Band.Personalize.Common.Test"
-        Square150x150Logo="Assets\Square150x150Logo.png"
-        Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="Band.Personalize.Common.Test"
-        BackgroundColor="transparent">
-        <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
-        <uap:SplashScreen Image="Assets\SplashScreen.png" />
-      </uap:VisualElements>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClient" />
-  </Capabilities> 
-</Package>
