--- src/Band.Personalize/Band.Personalize.Common.Test/UnitTestApp.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="Band.Personalize.Common.Test.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Band.Personalize.Common.Test">
-
-</Application>
