+++ src/Band.Personalize/Band.Personalize.Model.Implementation/project.json
