+++ src/Band.Personalize/Band.Personalize.Model.Library/project.json
