+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/LockScreenLogo.scale-200.png
