+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/SplashScreen.scale-200.png
