+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/Square150x150Logo.scale-200.png
