+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/Square44x44Logo.scale-200.png
