+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
