+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/StoreLogo.png
