+++ src/Band.Personalize/Band.Personalize.Model.Test/Assets/Wide310x150Logo.scale-200.png
