+++ src/Band.Personalize/Band.Personalize.Model.Test/Properties/UnitTestApp.rd.xml
