+++ src/Band.Personalize/Band.Personalize.Model.Test/UnitTestApp.xaml
@@ -1,0 +1,7 @@
+﻿<Application
+    x:Class="Band.Personalize.Model.Test.App"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Band.Personalize.Model.Test">
+
+</Application>
