+++ src/Band.Personalize/Band.Personalize.Model.Test/project.json
